all: diffing for arbitrary slices
This is the initial implementation for diffing two
slices of arbitrary type. This is the minimal
implementation that I need for some not yet
published projects I am working on and might
already be useful for other use cases as well.

That said, there are number of improvements and
additions that I would like to implement and the
API might change as well.