all: remove anchoring heuristic option
The anchoring heuristic is automatically used for larger diffs. For smaller
diffs, the result is either the same or slightly worse (in terms of number of
edits). It's possible there is a subjective improvement in diff qualtiy, but
that's very hard to evaluate and we already have the identation heuristic for
that. Furthermore, it's unclear what the overall user experince should be with
regards to knobs like this one. So I am going to remove it for now to think more
about this.